Fix no_data in failed job and error count monitors
Apply rollup function, because, as responded by DataDog support:

The issue on the initial monitors was that during evaluation, the timeshift
that necessarily operates on a change monitor, couldn't find datapoints that
were there if the datapoints were not submitted at the exact timeshift
timestamp of the monitor:

    Let's say the timeshift was looking for a datapoint at 8:34:25 am , and
you had one at 8:34:23 and one at 8:34:29, you'd artificially get a no data
result.

Appending the rollup(avg,30) function is correcting that in the way that it
will average the points buckets over a 30 sec timeframe, so you'd always
have data.