show-cf-memory-usage: Link to memory graph
Make it easier for people to know what they should compare the value of
"memory required" against. I've previously run the script but not verified
whether we are meeting ADR017 because it was too much effort to look up the
correct graphs.

The link shows the sum of memory provisioned to all cells in an environment.
I chose to use DataDog instead of our internal Graphite/Grafana because I
find it easier to access; I'm normally logged in and have my personal
credentials stored in a password manager, unlike the shared credentials for
environments.

I've built the URL with the `URI` module to make it easier to read and
maintain in code. A handful of query params have been omitted, such as
from/to timestamps and graph size, of which DataDog will redirect the
browser to a URL with the default values for these things.

I've had to add `check-env` to the Make target because we need the
`DEPLOY_ENV` environment variable to identify the correct environment in
DataDog. This does an unnecessary AWS credentials check, but I'm not going
to worry about that for now.

With the change to `check-env` in the preceding commit this also fixes a bug
whereby `API_ENDPOINT` wasn't populated and you would always get the warning
about targeting the right environment if you ran `make show-cf-memory-usage`
without an environment target.