make check-env: Check that env target has been run
Add a test to the Make `check-env` target which uses an additional
environment variable to confirm whether one of the environment targets
(dev/ci/staging/prod) have been run. The implementation seems a bit
kludgey, but I'm not aware of a Make-way of doing this other than string
parsing `MAKECMDGOALS`.

We currently kind-of rely on `DEPLOY_ENV` for this, but it doesn't guarantee
that an environment target has been run, because the environment variable
comes from outside the Makefile for dev environments.

I became aware of this while working on the `show-cf-memory-usage` target.
It uses `SYSTEM_DNS_ZONE_NAME`, which is set in the environment targets, but
it is possible to run it without as `make show-cf-memory-usage` and results
in an incorrect `API_ENDPOINT`. Adding `check-env` alone wasn't sufficient
for the reasons described above.