Increase default number of cells
Due to performance problems during acceptance tests execution we decided
to increase default number of cells to cope with
the load and allow tests to finish undisrupted. During
investigation with Pivotal we have discovered that metron is dropping
messages and UDP packed are being dropped by kernel as the load on cells
was extremely high. Also using all three AZs gives us some parity with
production.