Add tests for environment-specific cloud config
We want to configure a different disk size for elasticsearch_master
VMs in production only. Currently we only have environment-specific
configuration for our manifest, which has now been separated in the
unit tests. This adds the test, which should currently fail, and renames
a helper method to be consistent for both manifest and cloud config.