Increase elasticsearch data storage only in prod
There's currently an open alert for elasticsearch disk space, which is ~80%
used. Usage is not increasing significantly.

We don't want to manage two environment-specific configuration files, so
we move the current one up a directory and use it for both the manifest and
the cloud config.

Calculations for new disk space:

Current usage:               307200 * 0.784 * 3 = 722535
If one node dies:            722535 / 2 = 361268
To stay under 70% usage:     361268 / 70 * 100 = 516097 = 504GB
Round to 500GB for tidiness: 500 * 1024 = 512000

Current usage is based on current disk size in MB, the percentage usage indicated
by Datadog (78.4% on average) and the number of nodes. 500GB should keep us
happy even if we have to temporarily do without one node, whether it be due to
outage or deployment problems.