Use `aws_route` resources for internet route table
Terraform doesn't allow mixing routes defined inline in an
`aws_route_table` resource with separate `aws_route` resources[1]. We
need to start using the latter to add routes for the VPC peering, so we
need to migrate this.

This is non-straightforward because if we remove the inline `route`
statement, terraform wom't remove the routes, which means that when it
attempts to create the replacement `aws_route` resource, it gets a
conflict. This will be solved by modifying the tfstate file before
running terraform (in a separate commit so it can be reverted)

[1]https://www.terraform.io/docs/providers/aws/r/route_table.html