Give `scim.invite` scope to our admin users
Allows the global `admin` user (used in our pipeline and in emergencies if
SSO is broken) and our individual admin accounts to generate new account
invites via UAA:

- https://docs.cloudfoundry.org/api/uaa/#invite-users

This will be used to replace the password generation in `create-user.sh`.

The scope must be added to the `cf` client in order for it to be present in
tokens for users that have been given the UAA scope. This means we can send
invites with our normal CF user accounts, rather than using the global
`secrets.uaa_admin_client_secret` credentials. The scope will not be present
in tokens for normal users that do not have the UAA scope.