create-user.sh: Replace passwords with invites
Use UAA's builtin mechanism for generating (but not sending) invites for new
accounts. A link with a one-time use code allows the user to set their own
password. The code is invalidated after first use. This is preferable to
generating passwords because we don't need to send them in plaintext email
and we don't need to recommend that people change their password after the
first login.

We need to tell CF to create a user after generating the invite otherwise we
don't have an account to associate organisation and space membership to
until they set their password.

I have managed to resist rewriting this script into Go or Ruby because I
didn't want to change too much at once. However I'd strongly recommend
rewriting it in future with a language that has real libraries for
HTTP/JSON or CF/UAA, error handling, and tests.

In the meantime I've used `jq` to parse the responses and added some
defensive error checking, such as checking the length of response arrays
before indexing them, so that we can return a meaningful error and dump the
original HTTP response. This is because `var=$(jq -er '.var')` will cause
the script to abort but won't give any error message.

I've used standard `curl` to access UAA because `cf-uaac` didn't provide any
advantage; it doesn't seem to support importing a token so you'd need to
re-authenticate, it doesn't have a verb for the invite endpoint, and it's
`uaac curl` command doesn't check response codes or persist the disable TLS
verify option.

I'm re-using the oAuth token from the `cf` CLI so that we don't need to
re-authenticate (including SSO users). The query for `/v2/info` to get the
UAA endpoint should trigger a refresh of the token if it's expired. The CLI
doesn't have a verb to get the TLS verify status so we need to scrape it
from the config.

It's worth noting that the `redirect_uri` param doesn't appear to be used
for users with UAA as their `origin`. I believe it's only used for external
auth providers, but it's still required by the API, so I'm providing it as
an empty string.

The email warns that the invite link will only work once. Hopefully this
should prevent people from opening it before they are ready to set a
password. There is still a chance that it might be invalidated by automated
email scanners that follow links, but we'll have to see how often that
happens before taking action.

I've added a link to our "choosing passwords" documentation in the email
because I'm removing it from alphagov/paas-tech-docs where we previously
told users to change their passwords after first login.