Add empty UAA client with links to docs
When a user accepts an account invite and successfully sets their password
through the UAA login web interface they are redirected to a homepage that
says "Where to?" with no other text or links. This is likely to confuse our
new users.

The only way to customise the homepage content is to add attributes to UAA
clients which makes boxes appear with names, links, and optionally an image.

I'm using a new client, with no grants or scopes, to add a link to our "get
started" documentation. The image is a screenshot of the words "Get started"
from the documentation. I tried using a scalar block but UAA refused to
start, so I've used a single long string that is stored in a separate file
to make the manifest easier to read.

The other option was to set `home_redirect` which immediately redirects user
to another location. But I think that would have also been confusing for
people that have just entered their password.