create-user.sh: Re-purpose -r to resend invite
Previously the `-r` or `--reset-user` argument would delete the user and
then recreate them with a new password. This seems undesirable in most cases
because it will lose their organisation and space associations.

With the introduction of invites in the preceding commit we need a way to
re-send invites for existing users. This is particularly pertinent because
the invite links can only be used once, so users may invalidate them without
meaning to.

It didn't make sense to duplicate the invite functionality from
`create-user.sh` to `rotate-user-password.sh`, so I've re-purposed the
former and deprecated the latter. If a user tries to login while they have
an invite pending they will get this error message from the CLI:

    Authenticating...
    Server error, status code: 403, error code: access_denied, message: Account not verified

    Password> %