NGINX in front of UAA
Nginx is used to block /create_account and /create_account.do endpoints
of uaa. It is started on 8090 port and proxies the rest of the
traffic to localhost:8080 (UAA) preserving Host and X-Forwarded-For
headers.

The ELB maintains keepalive connections to UAA. We mimic that behaviour
by using the same keepalive timeout (20s) on the Nginx side, and using
keepalive connections from Nginx to UAA.

The logs are sent to syslog and can be found in logsearch.