Remove the API whitelist
This duplicates the way we have set up access for `web_access_cidrs` for
the API, removing the current set of whitelisted addresses and replacing
it with 0.0.0.0/0. It also renames the configuration variable from
`tenant_cidrs` to `api_access_cidrs` in order to bring it in line with
`web_access_cidrs` and make the new purpose clearer.

🐝