Add deploy_env tag to CF RDS
This tag will propagate through to the metrics that we get from the RDS
integration in DataDog. We can then use it to restrict monitors to a single
environment, rather than every environment in the AWS account (dev and CI
host more than one).

This isn't strictly necessary because the CF instance isn't a t2 that we need
to monitor. But it would be nice to be consistent with the BOSH instance.