Set MTU for Garden containers back to 1500
garden-runc 1.3.0, which was included by the CF upgrade in 46019be, changed
the MTU of Guardian containers to match the host's external network
interface instead of always being 1500:

- cloudfoundry/guardian@af8d612c728a2b2dd59fbd7feabd51b24e7ba132
- cloudfoundry/garden-runc-release@a9b22fd708203f304ab9837cace76d317125b999

For our cells on AWS this meant a change from 1500 to 9001:

- http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/network_mtu.html#jumbo_frame_instances

This didn't work well with IPsec, which we use to secure web traffic between
the routers and cells. Requests which generated responses larger than approx
9000 bytes were too large after the IPsec headers were added and never made
it back to the router.

It might have been nice to set this to a value a *bit* smaller than 9000, to
allow for the IPsec overhead, but we can't guarantee the MTU value of the
host interface because it gets set by AWS.

Ideally PMTU should mean that we wouldn't have to do this. We think that the
ICMP unreachable messages might not be getting back to the container (or
host?) correctly. It might be that we can revert this commit in future after
some more investigation.