Temporarily disable IPsec
Yesterday we manually turned this off on the cells and routers in response
to an incident. We want to codify this change, so that we can unblock other
deployments, and then the correct fix will follow soon after.

When IPsec is re-enabled we will need to change the level to `use` and then
do a subsequent deployment with `require`.