Deploy CDN broker VMs via BOSH
Add the jobs, properties, UAA permissions, ELB, security group for BOSH to
deploy the CDN broker VMs and configure them.

The CDN broker password (used for basic auth between the Cloud Controller and
the broker) and the broker client UAA password have been put in the secrets keep
file, so they will not be rotated.