Unset require_full_window for CPU credit monitors
The RDS monitor added in d8cfb6d is reporting "no data" in CI despite being
able to find metrics. This might be because the RDS metrics have a
resolution of 5mins compared to 2-3mins for EC2 metrics. We tested disabling
`require_full_window` and the monitor changed to "ok" state.

Although it's seemingly not affected at the moment, I'm applying the same to
the EC2 monitor to be safe. This still seems better than increasing the
window as described in 1f33b77. The other option would be to set "delay
evaluation" but that doesn't seem possible using the Terraform provider or
API.