Add CDN broker RDS instance
The CDN broker requires a database to maintain its state. We create a
dedicated RDS instance. It's separate from the CF instance as the broker
is not a core feature and can be enabled or not.

The CDN master DB password has been put in the secrets keep file, so it
will not be rotated.