Reduce false alarms due to etcd monitor
The "etcd IsLeader count is exactly one" monitor is triggered every time
the leader count is different than one. However, this happens frequently
for a few seconds because of deployments or normal etcd re-elections and
this creates a lot of false alarms.

Now it will alarm if the count is continuously different than 1 on a 1
minute period. This is a short interval but this metric arrives every
20-30s so this should be enough.