Unify generation of test configuration
We used to rely on the templates from cf-release to generate the
config.json files but this caused us issues like:
https://github.com/cloudfoundry/cf-release/issues/1197

We now take the same task as custom acceptance tests as it doesn't
depend on cf-release, make it generic and reuse it for all the tests.

Some properties had a default value in spec files or were changed in the
template. We now set the values explicitely inside the manifest
properties.