Fix create-user to work with cf-cli 6.26
With version 6.26 of the cli, the output from `cf target` has changed
slightly. The fieldname is now "user" (lowercase) instead of "User".
This was breaking the script.

Instead of just changing this, I've used the `tolower` awk function (as
described here[1]) so that it remains compatible with older versions of
the CLI.

[1]https://unix.stackexchange.com/questions/273624/case-insensitive-search-in-awk