Increase `go test` timeout for availability-tests
We increased the timeout of the long running `Eventually` block of the
`availability-tests` in 510471b but overlooked that we also have to provide
a timeout to `go test`.

I considered adding a comment to point people at one-or-the-other but it
seems unlikely that we'll need to change these values again.