Open web access security group
Web access to the platform (via router ELB then haproxy/gorouter) was
open in staging but closed in other environments. This does not protect
from a security risk, it requires keeping a list of IPs up-to-date and
makes testing cumbersome. For example to test cloudfront, the security
group has to be manually edited to open to the world.

We now open it to the world in all environments and remove the
web_access_cidrs list.