Increase size + complexity of healthcheck response
The existing response body wasn't big enough to trigger the MTU bug that was
fixed by 4dd3c3f. By increasing the size and complexity of the response body
we hope to catch more edge cases like that bug, both in our pipeline tests
which are run after a deployment and our Pingdom checks which run all of the
time.

I've used the book "Alice's Adventures in Wonderland" because:

- at 220kB (on disk) it represents a reasonable sized website
- at 68kB (on disk after gzip) it is guaranteed to still be larger than our
  MTU size of 9kB
- the content is free to use from Project Gutenberg
- it contains real HTML

I replaced `index.html` with the file `wget`ed from:

- https://www.gutenberg.org/files/11/11-h/11-h.htm

I considered using only some of the text, but it was easier to use the whole
thing, especially as it has a license at the bottom. I've left the Windows
line endings from the original file as they shouldn't affect us.

Some custom acceptance tests are included which querying the healthcheck app
as deployed to the platform by our pipeline. They check that the response
size meets the criteria described above. This should give us an early
warning of the functionality or intention of the Pingdom checks changing.