Refactor cf-deploy aggregation
Split an aggregate of 13 tasks into two groups. This is because of
concerns regarding the load on Concourse of having so many tasks
running in parallel. The split is between tasks relating to the operation
of CloudFoundry, and those which are more to do with the monitoring of
the platform. The one exception is the registration of the Compose
broker, which has to take place after the deployment of the Compose
broker application.