Allow re-inviting user without assigning to org/space
When re-inviting a user (eg because they've forgotten their password),
we don't need to specify an org because they'll already be assigned to
the orgs/spaces then need access to. In this case it's desirable to not
touch the orgs/spaces.

This therefore allows this script to run without an org specified when
re-inviting a user.