Better concourse thrashing alert monitoring
This introduces a new monitor using the new metric tracking the number
of btrfs processes running at any one time. When concourse starts to
fail the number of processes spikes quickly and drastically, so this
should allow us to quickly spot when concourse needs shaking before the
load average catches up.

Numbers:
* Processes matching "btrfs" on an idle concourse: 23
* Processes matching "btrfs" on an active concourse: 23
* Processes matching "btrfs" on a currently broken concourse: 242

Also at @alext's suggestion this tweaks the existing CPU load monitor so
that it will trigger within a minute of the load average spiking over
the limit.