Use European keyservers in git-resource GPG check
We have seen intermittent problems with requests to keyservers timing out
and failing the `pipeline-lock` job in our pipeline:

    gpg: requesting key 6DCAB6FA from hkp server keys.gnupg.net
    gpg: keyserver timed out
    gpg: keyserver receive failed: Keyserver error

The default of `keys.gnupg.net` is a CNAME for `pool.sks-keyservers.net`
which is described at:

- https://sks-keyservers.net/overview-of-pools.php

Try using the European pool, instead of the default pool, in the hope that
the lower latency will make them more reliable. If this doesn't work and we
think it's still a problem them we could consider inline all of the public
key contents in the future.