Add custom acceptance test for Compose broker
This is the starting point for our tests. It only tests the service
is registered and can be made available to users via the marketplace.
Tests that interact directly with the broker have not been implemented,
as the interaction between the Cloud Controller is already proven by
being able to register it in the first place (to register the broker
the Cloud Controller has to hit the /v2/catalog endpoint to fetch the
catalog).

We only want to enable the service for a specific org and this operation
is not idempotent[1], so we have to use the BeforeSuite hook to enable
it.

[1] https://github.com/cloudfoundry/cli/issues/939