Add the parser for uaa_nginx_access
The UAA NGINX access log comes in standard format. With the following
parser, we'll be able to extract the details of each log which will also
allow us to perform advanced query with Kibana.

We could have used the original httpd log grok pattern from:
https://github.com/logstash-plugins/logstash-patterns-core/blob/master/patterns/httpd
But we want to prefix the fields with [nginx]