Pin all governmentpaas containers to a version
So that we can control how containers are updated between environments and
prevent forwards/backwards incompatibilities with code in the pipeline.

This will also prevent Concourse from re-downloading all of the containers
each time we merge a change to alphagov/paas-docker-cloudfoundry-tools
because Docker Hub rebuilds everything.

Using the functionality added in:

- alphagov/paas-docker-cloudfoundry-tools#93
- alphagov/paas-semver-resource#5
- alphagov/paas-s3-resource#5

The `run-bosh-cli` task uses a different format because the "bug" of needing
to manually choose the `check` or `task` container hasn't been fixed yet.
There's more information in:

- https://webcache.googleusercontent.com/search?q=cache:9l5PuEzf5e4J:https://github.com/concourse/fly/issues/98+&cd=4&hl=en&ct=clnk&gl=uk
- https://github.com/concourse/concourse/issues/629#issuecomment-299445449