Increase thresholds for cc.log_count.error alert
Somethings appears to have changed since the CF v257 upgrade and the
acceptance tests are now alerting me via PagerDuty on every successful run.

I've tracked this down to a log entry and found the corresponding API
requests in the test artefacts:

    Stager error: Unsupported buildpack type: 'CATS-5-BPK-c816e215-00d4-4b56-7'

However I'm not able to find a fix yet and the alerts are very distracting.
I also think that this threshold is too low, so it seems safe to raise it.