avail-tests: Print the error and continue
Print the error but do not stop the tests.

The job availability-tests/470[1] has failed. We think it actually
shouldn't matter to us that much, if there is a one of connection issue
with the concourse API. We don't mean to get rid of the error
completely, therefore we're still printing it, allowing the script to
continue it's loop and hoping that the issue has resolved it self in the
mean time.

[1] https://deployer.cloud.service.gov.uk/teams/main/pipelines/create-cloudfoundry/jobs/availability-tests/builds/470