Increase the cell disk size
As we grow, we're slowly running out of disk space. We've reached a
little over 75% on two of our cells in production environment, according
to DataDog.

We'd like to double the disk space, for the time to stay safe and
alerted when other warnings hit production.

We have decided, to increase the disk space across all environments.