Custom acceptance tests for HTTP(s) edge cases
We want our acceptance tests to tell us about regressions similar to the
MTU response size incident. This commit implements a set of HTTP tests
for edge cases that will try to replicate weird requests to apps we host
on our platform. It also adds a HTTP tester app with multiple endpoints
that can return headers or query strings in the body so tests can make
assertions.

This set of tests replaces response_size_test and print_request_headers
app that we used before.

List of tests:
- large request bodies
- large request headers
- large response headers
- egress connectivity to a service outside the PaaS
- app-to-app connectivity between two apps on the PaaS
- slow response
- GET with many parameters
- GET with UTF8 parameters?
- GET with up to 2000 characters