Reserve common hostnames on the apps domain
We do not want a tenant to be able to push an app using any of these
names, as it could be misleading to others[1].

There was a suggestion these hostnames should redirect to the product
page, but it was agreed to not add that complexity at this stage.
However, in case we do they have been placed in the same org as the
product page (govuk-paas), which seems to have been created manually
in the past.

The redundant `echo | ` has been removed because it does not appear
to do anything with our current usage of the CF CLI.

[1] https://www.pivotaltracker.com/story/show/145588187