Upgrade Go version used by the healthcheck app
Upgrade to Go 1.8 because that is the latest version supported by
the Go buildpack we are currently using[1]. The buildpack is Godep-aware
and will compile based on this version.

Files changed using `godep update -goversion`, which changes the files
based on what you have installed locally. Godep v79 is the version supported
by the buildpack[1].

The app is pushed using the blue-green-deployment plugin, which creates
a new app, therefore eliminating the need to restage when this is deployed.

[1] https://github.com/cloudfoundry/go-buildpack/releases/tag/v1.7.19