cflinuxfs2 CVE update
It was discovered that JasPer incorrectly handled certain malformed JPEG-2000 image files.
If a user or automated system using JasPer were tricked into opening a specially crafted image,
an attacker could exploit this to cause a denial of service or possibly
execute code with the privileges of the user invoking the program.

We have to update to cflinuxfs2 v 1.123.0+

https://www.cloudfoundry.org/usn-3295-1/