Use create_before_destroy lifecycle for security groups.
AWS does not allow changing the description for a secuirty group. As a
result terraform works around this by deleting and recreating the
security group. This fails because the security groups are in use. It
would also cause a short period of downtime.

To avoid this we use the create_before_destroy lifecycle option. This
instructs terraform to create a new security group, and update all the
dependencies to use it before deleting the old one. Using name_prefix is
required to allow terraform to generate non-conflicting names for these
groups.