Pin compose broker version
As we want to have control over versions we deploy and prevent
auto-updates we need to pin the broker version.

We've found previous method of version pining problematic in the past.
It won't work for new deployments (or newly pushed pipelines) if this commit isn't the HEAD of the branch this resource is pointing at.
The reason for this is that this feature can only select from versions that the resource has seen.
For a new deployment, this list will only include the HEAD commit, and not any previous ones.