Upgrade ipsec boshrelease
We've applied some logrotate fixes to the boshrelease, which will no
longer prevent logrotate from rotating the logs on ipsec.

Update by alext:

After further investigation, the errors that we were seeing were caused
by the way that logrotate was being invoked (`logrotate /etc/logrodate.d/vcap`
instead of `logrotate /etc/logrotate.conf`). There errors do not occur
when it's invoked in the same way as the cronjob.

The upstream PR to this release[1] has already been merged though, so
it's still worth updating the release to remain up to date.

[1]https://github.com/SAP/ipsec-release/pull/8