Remove comment regarding Diego 1.14
As stated in the release notes for garden-runc-release 1.5.0[1], the
`garden.cleanup_process_dirs_on_wait` needed to be set to true to retain
old behaviour for managing process state for containers. This comment was
added presumably because Diego 1.14 was to set this value to true, but
the release notes[2], say only the manifest generation scripts were changed,
which we do not use.

The default value is currently `false` in the job's spec[3], so we should
continue to set this in the manifest.

[1] https://github.com/cloudfoundry/garden-runc-release/releases/tag/v1.5.0
[2] https://github.com/cloudfoundry/diego-release/releases/tag/v1.14.0
[3] https://github.com/cloudfoundry/garden-runc-release/blob/v1.8.0/jobs/garden/spec#L158-L160