Use anomaly detection for cc errors.
Current alert triggers always if the change if more that 15 errors per half an hour which feels not very suitable for prod. Also we've got false positives when metric is reset after CC redeployment.
The anomaly detection method works for UAA and it seems to be correctly detecting anomalies
while allowing for counter resets.