Added MongoDB functionality to Healthcheck
The RDS acceptance tests deploy the Healthcheck example application,
bind it to RDS instances and test that basic read-write functionality
works for MySQL and Postgres.

We have extended this Healthcheck application with similar usage of
Compose's MongoDB. We plan to now create new acceptance tests of the
Compose Broker that use this updated Healthcheck app to ensure that
the MongoDB support is usable.

The Godeps.json now contains a fully-resolved ImportPath rather than
the previous `healthcheck` ImportPath. We're not sure why it changed
but we ran `godep save .` when paas-cf was checked out within our
GOPATH.