Acceptance testing of Compose's MongoDB
We have added acceptance tests for the MongoDB functionality available
through the Compose Broker. These tests use the Healthcheck application
to perform read-write operations on a MongoDB service and check this
succeeds. We have previously extended the Healthcheck application to
perform these operations on MongoDB.

Much of the test code was adapted from `rds_broker_test.go`.
We didn't copy the `getRDSInstanceName` method which gets the
RDS DB instance identifier from the broker's last operation
output. We believe this was to identify RDS instances from test
runs. We have not included this because the Compose broker does
not give us anything comparable.