Patch cf-acceptance-tests
We have split the capi-release from the cf-release in order to mitigate
CVEs[1]. This means tests were failing because deleting apps using the V3
API is now an asynchronous request and responds with a `202 Accepted`, not
the `204 No Content`.

We can remove this once the capi-release and cf-release are back in sync.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/149220549