pre-deploy should fail if app-availability-tests never start
In the original version of the pre-deploy job there was no check to see
wether the healthcheck app was deployed or not. It was assumed that if
the app-availability-tests job could not be detected as running that it
probably wasn't deployed and so the pipeline should just continue.

This also meant that any other errors in this stage were silently
ignored and there was a situation where pre-deploy would pass OK even
though the app-availability-job had not begun yet, which _could_ lead
to incorrect statistics being gathered (as the healthcheck app might
start late in the process).

A method for checking if the healthcheck app is deployed does
exist now (deployed-healthcheck file in state bucket) and so the
pre-deploy stage should not be allowed to silently fail and should
always wait for the *-availability-tests jobs to begin.