Add CloudFoundary API monitoring framework
We want to be able to gather basic statistics on the availability of the
cloudfoundry API, specifically during deployment.

We want this to simulate real world usage for example "does `cf app
myapp` work"

This change adds a simple task running framework where:
  * each "task" is a func that gets given a cf client instances
  * each task should execute a set of api calls to simulate a given
scenaerio (ie `cf app myapp`)
  * each task will be executed repeatably at a given rate
  * statistics will be gathered grouped by task for
    - executions per second
    - error rate

It's a bit like a cloud foundry api specific version of vegeta. Sort of.